refactor: move wallet to its own folder
