+++ packages/server-wallet/src/wallet/index.ts
@@ -1,0 +1,2 @@
+export {Wallet} from './wallet';
+export * from './types';
